feat(scalars): detach string components from form
