+++ packages/design-system/src/scalars/components/fragments/text-field/index.ts
@@ -1,1 +1,2 @@
-export * from "./text-field.js";
+export { TextField, type TextFieldProps } from "./text-field.js";
+export { TextInput, type TextInputProps } from "./text-input.js";
