+++ packages/design-system/src/scalars/components/fragments/text-field/types.ts
@@ -1,0 +1,11 @@
+export interface CommonTextProps {
+  minLength?: number;
+  maxLength?: number;
+  pattern?: RegExp;
+  placeholder?: string;
+  trim?: boolean;
+  uppercase?: boolean;
+  lowercase?: boolean;
+  autoComplete?: boolean;
+  spellCheck?: boolean;
+}
