+++ packages/design-system/src/scalars/components/fragments/textarea-field/index.ts
@@ -1,1 +1,2 @@
-export * from "./textarea-field.js";
+export { TextareaField, type TextareaFieldProps } from "./textarea-field.js";
+export { Textarea, type TextareaProps } from "./textarea.js";
