+++ packages/design-system/src/scalars/components/string-field/string-field.tsx
@@ -3,10 +3,7 @@ import {
   TextField,
   type TextFieldProps,
 } from "../fragments/text-field/index.js";
-import {
-  TextareaField,
-  type TextareaProps,
-} from "../fragments/textarea-field/index.js";
+import { Textarea, type TextareaProps } from "../fragments/textarea-field/index.js";
 
 export interface StringFieldProps
   extends Omit<TextFieldProps, keyof TextareaProps>,
@@ -19,7 +16,7 @@ export const StringField = React.forwardRef<
   if (autoExpand || multiline) {
     // only textarea supports autoExpand and multiline
     return (
-      <TextareaField
+      <Textarea
         autoExpand={autoExpand}
         multiline={multiline}
         {...(props as TextareaProps)}
